LayoutTests/fast/webgpu/nocrash/fuzz-145583179-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: the end
CONSOLE MESSAGE: fuzz-145583179.html
